requirements.txt
imutils == 0.5.3
matplotlib == 3.3.2
numpy == 1.19.2
opencv_python == 4.4.0.44
py_sudoku == 1.0.1
scikit_image == 0.17.2
scikit_learn == 0.23.2
solver == 0.0.4
tensorflow_gpu == 2.3.1
